backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
psycopg2-binary==2.9.9
pymysql==1.1.1
email-validator==2.2.0
alembic==1.13.2
redis==5.0.8
celery==5.4.0
pgvector==0.3.6
sentence-transformers==3.3.1
einops==0.8.0
httpx==0.27.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic-settings==2.4.0
PyYAML==6.0.2
langgraph==0.3.5
langchain-community==0.3.20
langchain-ollama==0.2.3
ollama==0.4.7

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# New features
pyyaml==6.0.2
croniter==3.0.3
